Update components/incidents regardless of HARD/SOFT recovery
The original script ignored CRITICAL SOFT but we want that to result in a
component status change without creating an incident.

Previously, cachet_notify would only change a component's status if it hit CRITICAL HARD.

Now, we don't care if the component recovers SOFT or HARD.  Update the component
and/or incident either way.

Signed-off-by: David Galloway <dgallowa@redhat.com>